// src/api/__generated__/story/model/actorCollectionResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { ActorCollectionResponseBadge } from './actorCollectionResponseBadge';

/**
 * 数据列表
 */
export interface ActorCollectionResponse {
  id?: number;
  userId?: number;
  creatorName?: string;
  assetId?: string;
  name?: string;
  avatarUrl?: string;
  bio?: string;
  status?: string;
  auditReason?: string;
  nftMintAddress?: string;
  pricingMode?: string;
  badge?: ActorCollectionResponseBadge;
  totalSupply?: number;
  mintedSupply?: number;
  availableSupply?: number;
  initialPriceUsdc?: number;
  currentPriceUsdc?: number;
  trust?: number;
  initialPriceMultiplier?: number;
  computingPower?: number;
  nftChain?: string;
  nftTokenStandard?: string;
  nftTxHash?: string;
  completedViewCount?: number;
  heatValue?: number;
  miningCoefficient?: number;
  createdAt?: number;
  updatedAt?: number;
  version?: number;
}