// src/api/__generated__/story/model/mintActorNftRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { MintActorNftRequestPayMethod } from './mintActorNftRequestPayMethod';

export interface MintActorNftRequest {
  walletAddress: string;
  nftChain: string;
  nftTokenStandard: string;
  nftContractAddress: string;
  payMethod: MintActorNftRequestPayMethod;
}